requirements.txt
pandas==0.24.2
pytesseract==0.3.8
requests==2.22.0
beautifulsoup4==4.10.0
fitz==0.0.1.dev2
Pillow==8.3.2
